Improve logging of protocol buffers messages (#20)
Currently when a protocol buffers message is logged it is automatically
converted to JSON by the logging library. That doesn't use the same
fields names that are used in the API, and doesn't display the type of
the object. To improves that, so that the messages will be displayed
with the same filed names used by the API and with a `@type` field
containing the name of the type. For example, code like this:

```go
s.logger.InfoContext(
        ctx,
	"My object",
	slog.Any("object", myObject),
)
```

Will result in a log line like this if the object is a wrapped integer:

```json
{
  "time": "2025-03-31T11:00:17Z",
  "level": "INFO",
  "msg": "My object",
  "object": {
    "@type": "type.googleapis.com/google.protobuf.Int32Value",
    "value": 42
  }
}
```

Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>