A bundle of small changes (#18)
* Update clusterference inside mutate function

* Find namespace using a label selector

Since we generate the name with a random suffix it cannot be inferred from
the ClusterOrder. We need to find it using a label selector in order to
update status.clsuterReference appropriately.

* Only respond to ClusterOrders in a specific namespace

We want to only watch ClusterOrder in a specific namespace. This will
ultimately simplify some aspects of the design and will effectively enforce
unique names on ClusterOrders (which is not the case when they can be
create in arbitrary namespaces).

The solution here is sub-optimal and will be replaced in a subsequence
change.