Automatically create private clusters orders and clusters (#31)
Currently the objects containing private information about cluster
orders and clusters need to be created explicitly by the components that
use them. To simplify that this patch changes the server so that it
creates them automatically when the corresponding public object is
created. To do so it uses database triggers. The resulting SQL files are
much larger, so the patch also splits them into separate files for the
public and private schemas.

Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>