Don't initialize conditions for each reconcile
Currently we initialize all the conditions at the beginning of the
reconcile loop, and then we update them according to the state of the
order. This has the side effect of always updating the
`lastTranstionTimestamp` of the conditions, even when they haven't been
changed. To avoid that this patch changes the code so that only
initializes the conditions that haven't already been initialized.

Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>