Refactor network/node playbooks into single idempotent playbook
The previous attach/detach network/node playbooks could result in
unintended networks remaining on the node. The 'set_networks_for_node'
playbook will ensure that the node is configured with the exact
networks passed in.

Network UUIDs are used because the ESI 'esi node network list'
command returns a modified version of the network name.