Refactor infrastructure management roles
This commit reorganizes some of the infrastructure setup/teardown so that
(a) we can properly sequence setup/teardown of the port forwarding and dns
records with cluster teardown, (b) to resolve some errors caused by
sharing the same workflows for both create and delete operations, and (c)
to avoid some race conditions that caused agents to be erroneously removed
from a cluster immediately after they were added.

In particular, this moves the agent and network setup out of the
`hosted_cluster` role and into the `cluster_infra` role.