Add kind cluster as hub for integration tests
Currently the environment for integration tests doesn't have any hub, so
it isn't possible to test anything related to it, like the creation of
the Kubernetes `ClusterOrder` object. This patch changes the test suite
so that the kind cluster used to run the service is also used as a hub.

Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>