Wait for API server before starting reconcilers
Currently the reconcilers are started inmediately when the controller
process is started. If the API server hasn't started yet this will
result in a connection failure. The controller process tolerates that,
and will retry, but the retry is too slow for the timeouts of the
integration tests. To avoid that this patch changes the server so that
it waits for the API server to be ready before starting the controllers.
That way the integration tests will not depend on the retry interval.

Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>