Don't send hub Kubeconfig in notifications
Sending the Kubeconfig of hubs in notifications usually exceeds the size
max size of the PostgreSQL `NOTIFY ...` mechanism, which is 8000 bytes.
To avoid that this patch changes the generic server so that it removes
the Kubeconfig from notifications.

A more general way to solve this would be to store the payloads in a
separate table, and send only the identifier of the event in the payload
of the `NOTIFY` mechanism. We will do that later.

Related: https://github.com/innabox/fulfillment-service/issues/82
Related: https://www.postgresql.org/docs/current/sql-notify.html
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>