Add logger to recovery tests
Currently the logger used in the tests of the panic recovery interceptor
writes to thes standard output stream directly. This mixes the output
with the output of the `ginkgo` tool. To avoid that this patch uses a
logger that writes to the `GinkgoWriter`, like in other tests in the
project.

Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>