Lock failure must fail playbook (#116)
Previously, we treated failure to acquire a lease as a "successful
failure", but now that we're using a workflow template this doesn't make
sense: this causes the post-install playbook to execute after a lock
failure, which is generally the wrong thing to do.

With this commit, failure to acquire a lock results in an actual playbook
failure.