Use ownerReference to manage lease
Relying on a block/rescue/always construct to manage the lease was problematic for
at least two reasons:

- If a job was cancelled, this could prevent the lease cleanup code from
  running.
- Some Ansible errors are un-catchable, and cause Ansible to abort
  immediately rather than executing the `rescue` or `always` blocks [1]

In either case, the job would exit without cleaning up the lease, which
would in turn prevent any future job for the same target cluster from
successfully executing.

We resolve this by replacing the block/rescue/always construct with an
`ownerReference` [2] on the lease. This causes Kubernetes to remove the
lease for us automatically when the pod exits.

Because ownerReferences cannot cross namespaces, this requires creating the
Lease in the same namespaces as the AAP job, rather than in the "cluster
working namespace".

[1]: https://github.com/innabox/cloudkit-aap/issues/127
[2]: https://kubernetes.io/docs/concepts/overview/working-with-objects/owners-dependents/

Closes: #130