Support retries when acquiring a lease
In #142, we converted the manage_agents role to using the new
cloudkit.service.lease role, but while you can loop over an `include_role`
task with `loop`, you cannot use a until/retries loop.

This commit moves the retry loop into the lease role, with a default
"retries" value of 0, and adds a new "lease_retries" parameter to the role.