Run post-delete job to report status
Replace the post-delete job with a workflow that includes success and
failure nodes to update the clusterorder manifest with information about
the success or failure of the delete workflow.

We currently report status via annotations; ideally we would be updating
status conditions on the resource. See [1].

We want to report status for the create workflow as well, but this will be
implemented in a future change [2].

[1]: https://github.com/innabox/issues/issues/310
[2]: https://github.com/innabox/issues/issues/311